+++ CHANGELOG.md
@@ -41,6 +41,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
   instrumentation for the `github.com/syndtr/goleveldb/leveldb`
   package. (#186)
 - Add the
+  `github.com/signalfx/splunk-otel-go/instrumentation/github.com/tidwall/buntdb/splunkbuntdb`
+  instrumentation for the `github.com/tidwall/buntdb`
+  package. (#228)
+- Add the
   `github.com/signalfx/splunk-otel-go/instrumentation/k8s.io/client-go/splunkclient-go`
   instrumentation for the `k8s.io/client-go` package. (#224)
 - Add the